添加了很多检查位置编码问题的调试代码

+++ segmentation.py
@@ -110,7 +110,7 @@ trainer = Trainer.from_argparse_args(
     devices=1,
     auto_select_gpus=True,
     gradient_clip_val=1.0,
-    precision=16, # 启用16位混合精度
+    precision=32, # 启用16位混合精度
     amp_backend='native' # 指定使用PyTorch原生AMP后端
 )
